requirements.txt
# Framework principal
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Banco de dados e ORM
sqlalchemy==2.0.23
alembic==1.12.1

# Autenticação e segurança
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Ambiente e configurações
python-dotenv==1.0.0

# Testes
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Validação de dados
pydantic==2.5.0
email-validator==2.1.0

# Banco de dados (escolha um)
# Para SQLite (já vem com Python)
# Para PostgreSQL (opcional)
psycopg2-binary==2.9.9

# Para MySQL (opcional)
# mysqlclient==2.2.0